moved EdcScopeToCriterionTransformer to identity-hub-core
